requirements.txt
openai
requests
python-dotenv
pip-system-certs